+++ src/rdkafka_metadata.c
@@ -966,7 +966,8 @@ static void rd_kafka_metadata_leader_query_tmr_cb (rd_kafka_timers_t *rkts,
                     rk->rk_conf.metadata_refresh_interval_ms)
                         rd_kafka_timer_stop(rkts, rtmr, 1/*lock*/);
                 else
-                        rd_kafka_timer_backoff(rkts, rtmr, rtmr->rtmr_interval);
+                        rd_kafka_timer_backoff(rkts, rtmr,
+                                               (int)rtmr->rtmr_interval);
         }
 
         rd_kafka_wrunlock(rk);
